Extract number-formatters from misc
