+++ freqtrade/commands/optimize_commands.py
@@ -5,7 +5,7 @@ from freqtrade import constants
 from freqtrade.configuration import setup_utils_configuration
 from freqtrade.enums import RunMode
 from freqtrade.exceptions import OperationalException
-from freqtrade.misc import round_coin_value
+from freqtrade.util import round_coin_value
 
 
 logger = logging.getLogger(__name__)
