+++ freqtrade/util/__init__.py
@@ -1,6 +1,7 @@
 from freqtrade.util.datetime_helpers import (dt_floor_day, dt_from_ts, dt_humanize, dt_now, dt_ts,
                                              dt_ts_def, dt_utc, format_date, format_ms_time,
                                              shorten_date)
+from freqtrade.util.formatters import decimals_per_coin, round_coin_value, round_value
 from freqtrade.util.ft_precise import FtPrecise
 from freqtrade.util.periodic_cache import PeriodicCache
 from freqtrade.util.template_renderer import render_template, render_template_with_fallback  # noqa
@@ -19,4 +20,7 @@ __all__ = [
     'FtPrecise',
     'PeriodicCache',
     'shorten_date',
+    'decimals_per_coin',
+    'round_value',
+    'round_coin_value',
 ]
